Calculate network RX and TX without user config
Previously users would need to set the nework device name, with defaults that changed over time based on ethernet (eth0) or wireless (wlan0) and some other values (e.g. enp0s31f6). The script would then use this single network_name value to determine the amount of traffic received and transmitted.

This change removes the need to configure the device name, and instead checks all links, excluding those named "lo" for loopback, and then sums the values for RX and TX before converting to human readable format.

This relies on an assumption that the loopback device will be named "lo". While this is fairly safe, its not guaranteed as any distro could in the future use a different name.  If this arises, consider doing comparisons such as any of the following as alternatives for filtering
- `select(.link_type!="loopback")`
- `select(.link_type=="ether")`